have the buttons updating correctly
given that i dont need to save every time i created a class that saves  the first index everything is initilized as, using the array index. Then when it is changed i rotate by just adjusting the current index int up or down depending on the rotation. Always keeping track of the place to save it because of the array. So when saved the string will send each buttton to the intial start place, rather than its current position